Move 51destruct to existing 39objectdestructuring
