+++ tests/compile-test/lang-test0/pxt.json
@@ -53,7 +53,6 @@
         "48instanceof.ts",
         "49unicode.ts",
         "50indexedtypes.ts",
-        "51destruct.ts",
         "99final.ts"
     ],
     "public": true,
